split index into different files
