+++ test/protocol/src/helpers.ts
@@ -1,0 +1,5 @@
+import { URI } from 'vscode-uri'
+
+export function toUri(path: string) {
+  return URI.file(path).toString()
+}
